feat: Implementar geração de dados simulados e salvar em CSV
- Adicionada função `gerar_dados_simulados` para criar dados de telemetria em formato consolidado.
- Simulação de eventos como tempestades e suas influências na geração de energia e condições ambientais.
- Atualização da lógica de sensores e armazenamento de dados em um arquivo CSV.
- Remoção do arquivo `teste.txt` que não era mais necessário.
- Alteração das classes de sistemas para aceitar valores de geração como float.
- Correção na lógica de alertas para garantir que a severidade seja convertida corretamente.
- Criação da classe `MotorDeRegras` para gerenciar a lógica de análise da missão.
- Implementação da ingestão de dados do CSV no sistema principal, com atualizações nos sistemas e sensores.

+++ src/modules/modulos.py
@@ -183,12 +183,12 @@ class Sistema:
 
 class SistemaGeracaoSolar(Sistema):
     '''Painéis Solares.'''
-    def __init__(self, nome: str, capacidade_max_geracao: int, geracao_atual: int):
+    def __init__(self, nome: str, capacidade_max_geracao: int, geracao_atual: float):
         super().__init__(nome, capacidade_max_geracao, geracao_atual,consumo = 0, capacidade_max_armazenamento = 0)
 
 class SistemaGeracaoEolica(Sistema):
     '''Turbinas Eolicas.'''
-    def __init__(self, nome: str, capacidade_max_geracao: int, geracao_atual: int):
+    def __init__(self, nome: str, capacidade_max_geracao: int, geracao_atual: float):
         super().__init__(nome, capacidade_max_geracao, geracao_atual,consumo = 0, capacidade_max_armazenamento = 0)
 
 class SistemaArmazenamentoEnergetico(Sistema):
